Fix double calcularPreu y editar nombre funcion

+++ src/main/java/org/JavaCar/Llogable.java
@@ -1,5 +1,5 @@
 package org.JavaCar;
 
 public interface Llogable {
-    int calcularPreu(int dies);
+    double calcularPreu(int dies);
 }
